packages/sdk/python/pyproject.toml
[tool.poetry]
name = "superglue-client"
version = "1.0.0"
description = "Python SDK for superglue - the open-source data integration framework"
authors = ["superglue <hi@superglue.cloud>"]
license = "MIT"
readme = "README.md"
homepage = "https://superglue.ai"
repository = "https://github.com/superglue-ai/superglue/tree/main/packages/sdk/python"
documentation = "https://docs.superglue.cloud"
keywords = ["api", "sdk", "data-integration", "etl", "superglue"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
    { include = "superglue_client" },
]
include = ["superglue_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
